Fix type annotations and add utils tests
Type annotation fixes (all ty errors resolved):
- bilinear.py: Convert jnp.sqrt result to float
- gradients.py: Widen scale param to float|Array, fix return types
- hopfield.py: Fix beta type, modern_capacity return, use jnp.bool_
- masking.py: Use jnp.bool_ instead of bool for dtype
- multihead.py: Use Any for mixed return type dict
- utils.py: Import Figure/SubFigure, fix return types

Add tests/test_utils.py (41 tests):
- Tests for all 8 visualization functions
- Edge cases for masks, attention weights, metrics
- Proper matplotlib Agg backend for CI

Coverage increased from 64% to 95%

+++ src/attn_tensors/bilinear.py
@@ -54,7 +54,7 @@ def scaled_euclidean_metric(d: int, scale: float | None = None) -> Array:
         Scaled identity matrix of shape (d, d)
     """
     if scale is None:
-        scale = 1.0 / jnp.sqrt(d)
+        scale = float(1.0 / jnp.sqrt(d))
     return scale * jnp.eye(d)
